requirements/test.txt
# Test dependencies go here.
-r base.txt


django-debug-toolbar==1.4
django-extensions==1.6.1

coverage==4.0.3
django_coverage_plugin==1.2.2
flake8==2.5.4
django-test-plus==1.0.12
factory_boy==2.6.1
django-webtest==1.7.8
mock==1.3.0

# pytest!
pytest-django==2.9.1
pytest-sugar==0.5.1

